+++ src/main/java/com/github/propheticeclipse/tensurastarlight/ability/skill/unique/EchoEnduresSkill.java
@@ -77,7 +77,7 @@ public class EchoEnduresSkill extends Skill {
 
     @Override
     public double getAcquiringMagiculeCost(ManasSkillInstance instance) {
-        return 100000;
+        return 75000;
     }
 
     @Override