+++ src/main/java/com/github/propheticeclipse/tensurastarlight/registry/StarlightEffects.java
@@ -1,6 +1,7 @@
 package com.github.propheticeclipse.tensurastarlight.registry;
 
 import com.github.propheticeclipse.tensurastarlight.effect.buff.EmberAttackImprovementEffect;
+import com.github.propheticeclipse.tensurastarlight.effect.buff.RemnantsOfAscensionArmorEffect;
 import com.github.propheticeclipse.tensurastarlight.effect.buff.ShapeForgottenArmorEffect;
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
@@ -16,10 +17,12 @@ public class StarlightEffects {
     private static final DeferredRegister<MobEffect> MOB_EFFECTS;
     public static final RegistrySupplier<MobEffect> EMBER_ATTACK_IMPROVEMENT;
     public static final RegistrySupplier<MobEffect> SHAPE_FORGOTTEN_ARMOR;
+    public static final RegistrySupplier<MobEffect> REMNANT_ASCENSION_ARMOR;
 
     static {
         MOB_EFFECTS = DeferredRegister.create("trstarlight", Registries.MOB_EFFECT);
         EMBER_ATTACK_IMPROVEMENT = MOB_EFFECTS.register("ember_attack_improvement", EmberAttackImprovementEffect::new);
         SHAPE_FORGOTTEN_ARMOR = MOB_EFFECTS.register("shape_forgotten_armor", ShapeForgottenArmorEffect::new);
+        REMNANT_ASCENSION_ARMOR = MOB_EFFECTS.register("remnant_ascension_armor", RemnantsOfAscensionArmorEffect::new);
     }
 }