+++ CHANGELOG.md
@@ -16,8 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unique Skill: [Shape Forgotten]
 - Unique Skill: [Current Unbroken]
 - Evolved Unique Skill: [Vestiges Of Eidolons]
+- Evolved Unique Skill: [Remnants Of Ascension]
 
 - Effect: [Ember Remains Buff]
 - Effect: [Shape Forgotten Buff]
+- Effect: [Remnants Of Ascension Buff]
 
 ### Changed
+
+- Several Nerfs/Reworks from the 1.19.2 version of Tensura: Starlight, some of which listed below:
+- - Ember Remains Buff no longer being based on the in combat effect from mysticism, but rather based on dealing damage.
+- - Light Remains can now be acquired with having both Light and Dark greater spirits, however it will NOT be refunded if obtained this way.
+- - General nerfs to damage output, armor effectiveness and similar.